Added dynamically adding contract instances and testing what happens when a contract instance crashes and is restarted in ContractModel.
