+++ nix/pkgs/haskell/materialized-darwin/.plan.nix/plutus-contract.nix
@@ -147,8 +147,10 @@
           "Plutus/Trace/Tag"
           ] ++ (pkgs.lib).optionals (!(compiler.isGhcjs && true || system.isGhcjs || system.isWindows)) [
           "Plutus/Contract/Test"
-          "Plutus/Contract/Test/ContractModel"
           "Plutus/Contract/Test/Coverage"
+          "Plutus/Contract/Test/ContractModel"
+          "Plutus/Contract/Test/ContractModel/Internal"
+          "Plutus/Contract/Test/ContractModel/CrashTolerance"
           ];
         hsSourceDirs = [ "src" ];
         };
