+++ plutus-contract/plutus-contract.cabal
@@ -157,8 +157,10 @@ library
     if !(impl(ghcjs) || os(ghcjs) || os(windows))
         exposed-modules:
             Plutus.Contract.Test
-            Plutus.Contract.Test.ContractModel
             Plutus.Contract.Test.Coverage
+            Plutus.Contract.Test.ContractModel
+            Plutus.Contract.Test.ContractModel.Internal
+            Plutus.Contract.Test.ContractModel.CrashTolerance
         build-depends:
             tasty -any,
             tasty-hunit -any,
