+++ plutus-contract/src/Plutus/Trace/Emulator.hs
@@ -17,6 +17,7 @@ An emulator trace is a contract trace that can be run in the Plutus emulator.
 module Plutus.Trace.Emulator(
     Emulator
     , EmulatorTrace
+    , EmulatorTraceNoStartContract
     , Wallet.Emulator.Stream.EmulatorErr(..)
     , Plutus.Trace.Emulator.Types.ContractHandle(..)
     , ContractInstanceTag
@@ -109,7 +110,7 @@ import Plutus.Trace.Effects.EmulatedWalletAPI (EmulatedWalletAPI, handleEmulated
 import Plutus.Trace.Effects.EmulatedWalletAPI qualified as EmulatedWalletAPI
 import Plutus.Trace.Effects.EmulatorControl (EmulatorControl, handleEmulatorControl)
 import Plutus.Trace.Effects.EmulatorControl qualified as EmulatorControl
-import Plutus.Trace.Effects.RunContract (RunContract, handleRunContract)
+import Plutus.Trace.Effects.RunContract (RunContract, StartContract, handleRunContract, handleStartContract)
 import Plutus.Trace.Effects.RunContract qualified as RunContract
 import Plutus.Trace.Effects.Waiting (Waiting, handleWaiting)
 import Plutus.Trace.Effects.Waiting qualified as Waiting
@@ -138,9 +139,21 @@ data PrintEffect r where
   PrintLn :: String -> PrintEffect ()
 makeEffect ''PrintEffect
 
-type EmulatorTrace a =
+type EmulatorTraceNoStartContract a =
         Eff
             '[ RunContract
+            , Assert
+            , Waiting
+            , EmulatorControl
+            , EmulatedWalletAPI
+            , LogMsg String
+            , Error EmulatorRuntimeError
+            ] a
+
+type EmulatorTrace a =
+        Eff
+            '[ StartContract
+            , RunContract
             , Assert
             , Waiting
             , EmulatorControl
@@ -171,6 +184,7 @@ handleEmulatorTrace slotCfg action = do
             . interpret (handleWaiting @_ @effs slotCfg)
             . interpret (handleAssert @_ @effs)
             . interpret (handleRunContract @_ @effs)
+            . interpret (handleStartContract @_ @effs)
             $ raiseEnd action
     void $ exit @effs @EmulatorMessage
