+++ plutus-contract/test/Spec/ErrorChecking.hs
@@ -48,23 +48,23 @@ tests = testGroup "error checking"
 
 -- | Normal failures should be allowed
 prop_FailFalse :: Property
-prop_FailFalse = checkErrorWhitelist handleSpecs defaultWhitelist (Actions [FailFalse])
+prop_FailFalse = checkErrorWhitelist defaultWhitelist (Actions [FailFalse])
 
 -- | Head Nil failure should not be allowed
 prop_FailHeadNil :: Property
-prop_FailHeadNil = checkErrorWhitelist handleSpecs defaultWhitelist (Actions [FailHeadNil])
+prop_FailHeadNil = checkErrorWhitelist defaultWhitelist (Actions [FailHeadNil])
 
 -- | Division by zero failure should not be allowed
 prop_DivZero :: Property
-prop_DivZero = checkErrorWhitelist handleSpecs defaultWhitelist (Actions [DivZero])
+prop_DivZero = checkErrorWhitelist defaultWhitelist (Actions [DivZero])
 
 -- | Division by zero failure should not be allowed (tracing before the failure).
 prop_DivZero_t :: Property
-prop_DivZero_t = checkErrorWhitelist handleSpecs defaultWhitelist (Actions [DivZero_t])
+prop_DivZero_t = checkErrorWhitelist defaultWhitelist (Actions [DivZero_t])
 
 -- | Successful validation should be allowed
 prop_Success :: Property
-prop_Success = checkErrorWhitelist handleSpecs defaultWhitelist (Actions [Success])
+prop_Success = checkErrorWhitelist defaultWhitelist (Actions [Success])
 
 -- | This QuickCheck model only provides an interface to the validators used in this
 -- test that are convenient for testing them in isolation.
@@ -103,6 +103,8 @@ instance ContractModel DummyModel where
 
   initialState = DummyModel
 
+  initialHandleSpecs = [ContractInstanceSpec (WalletKey w1) w1 contract]
+
   nextState _ = wait 2
 
   arbitraryAction _ = elements [FailFalse, FailHeadNil, DivZero, DivZero_t, Success]
@@ -117,9 +119,6 @@ type Schema = Endpoint "failFalse" ()
             .\/ Endpoint "divZero" ()
             .\/ Endpoint "divZero_t" ()
             .\/ Endpoint "success" ()
-
-handleSpecs :: [ContractInstanceSpec DummyModel]
-handleSpecs = [ContractInstanceSpec (WalletKey w1) w1 contract]
 
 -- | For each endpoint in the schema: pay to the corresponding validator
 -- and then spend that UTxO
