+++ plutus-use-cases/test/Spec/Auction.hs
@@ -194,6 +194,8 @@ instance ContractModel AuctionModel where
         , _phase      = NotStarted
         }
 
+    initialHandleSpecs = ContractInstanceSpec SellerH w1 seller : [ ContractInstanceSpec (BuyerH w) w (buyer threadToken) | w <- [w2, w3, w4] ]
+
     arbitraryAction s
         | p /= NotStarted =
             oneof [ WaitUntil . step <$> choose (1, 10 :: Integer)
@@ -275,18 +277,14 @@ instance ContractModel AuctionModel where
 
     monitoring _ _ = id
 
-delay :: Integer -> Trace.EmulatorTrace ()
+delay :: Integer -> Trace.EmulatorTraceNoStartContract ()
 delay n = void $ Trace.waitNSlots $ fromIntegral n
 
 prop_Auction :: Actions AuctionModel -> Property
 prop_Auction script =
-    propRunActionsWithOptions (set minLogLevel Info options) defaultCoverageOptions handleSpec
+    propRunActionsWithOptions (set minLogLevel Info options) defaultCoverageOptions
         (\ _ -> pure True)  -- TODO: check termination
         script
-
-handleSpec :: [ContractInstanceSpec AuctionModel]
-handleSpec = ContractInstanceSpec SellerH w1 seller :
-             [ ContractInstanceSpec (BuyerH w) w (buyer threadToken) | w <- [w2, w3, w4] ]
 
 finishAuction :: DL AuctionModel ()
 finishAuction = do
@@ -314,7 +312,7 @@ noLockProof = NoLockedFundsProof
       when (slot < 101) $ action $ WaitUntil 101
 
 prop_NoLockedFunds :: Property
-prop_NoLockedFunds = checkNoLockedFundsProof options handleSpec noLockProof
+prop_NoLockedFunds = checkNoLockedFundsProof options noLockProof
 
 tests :: TestTree
 tests =
